Add files via upload
Please note the original code in ipynb is different from the original script in  Python. But the original code in Python could not be run. 
After correcting the code in Python in accordance with ipynb, the Python script could be run. 

"""
Could not find compiler for platform Host: NOT_FOUND: could not find registered compiler for platform Host -- was support for that platform linked in?
	 [[{{node StatefulPartitionedCall}}]] [Op:__inference_train_function_25048]

"""